Уносим обработку выполнение из события функции в основной цикл выполнения программы

+++ MinesServer/GameShit/Entities/PEntity.cs
@@ -30,7 +30,7 @@ namespace MinesServer.GameShit.Entities
         public virtual int Health { get; set; }
         public virtual int MaxHealth { get; set; }
         public virtual int pause { get; set; }
-        public virtual double ServerPause { get; }
+        public virtual long ServerPause { get; }
 
         public virtual int cid { get; set; }
         public virtual int dir { get; set; }
